refactor: Divide into controller, domain, repository, service
