--- src/main/java/com/ssafy/todolist/TodoRepository.java
@@ -1,7 +1,0 @@
-package com.ssafy.todolist;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TodoRepository extends JpaRepository<Todo, Integer> {
-
-}
