+++ src/main/java/com/ssafy/todolist/repository/TodoRepository.java
@@ -1,0 +1,10 @@
+package com.ssafy.todolist.repository;
+
+import com.ssafy.todolist.domain.Todo;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TodoRepository extends JpaRepository<Todo, Integer> {
+
+}
